Drush backup script. Proposed solution to #12
Instead of having the script edit the crontab, I decided a better method would be having the ability to call the backup script from cron. You can pass it a directory to backup to, or it defaults to /vagrant/backups.

Format for the cron entry (for a backup at 2am each morning):
0 2 * * * /path/to/script/drush-backup.sh /optional/backup_dir/parameter